chore: remove unused import
